Drop search tables instead of migrating the data
