+++ core-bundle/src/Resources/config/migrations.yml
@@ -13,7 +13,6 @@ services:
         arguments:
             - '@database_connection'
 
-    Contao\CoreBundle\Migration\Version410\SearchWordsMigration:
+    Contao\CoreBundle\Migration\Version410\DropSearchMigration:
         arguments:
             - '@database_connection'
-            - '@contao.doctrine.schema_provider'
